added register method
