+++ models/User.js
@@ -1,0 +1,23 @@
+const mongoose = require('mongoose')
+const Schema = mongoose.Schema
+
+const UserSchema = new Schema(
+    {
+        name: {
+            type: String, 
+            required: true, 
+        }, 
+        email: {
+            type: String, 
+            required: true
+        }, 
+        password: {
+            type: String, 
+            required: true
+        }
+    }
+)
+
+const model = mongoose.model('UserSchema', UserSchema)
+
+module.exports = model
