+++ server.js
@@ -1,6 +1,9 @@
 const express = require('express')
 const mongoose = require('mongoose')
 const cors = require('cors')
+const Settings = require('./models/Settings')
+const User = require('./models/User')
+const bcrypt = require('bcryptjs')
 const PORT = 3004
 
 const app = express()
@@ -15,8 +18,6 @@ mongoose.connect("mongodb://127.0.0.1:27017/movie-game", {
 
     .then((console.log('Connected to database')))
     .catch(console.error)
-
-const Settings = require('./models/Settings')
 
 app.get('/settings', async (req, res) => {
     const settings = await Settings.find()
@@ -49,5 +50,30 @@ app.delete('/settings/delete/:id', async (req, res) => {
 
     res.json(result)
 })
+
+
+
+
+app.post('/register', async (req, res) => {
+    try {
+        const newPassword = await bcrypt.hash(req.body.password, 10)
+        await User.create({
+            name: req.body.name, 
+            email: req.body.email, 
+            password: newPassword
+        })
+        res.json({ status: 'ok' })
+    } catch(err) {
+        res.json({ status: 'error', error: 'Duplicate Email' })
+    }
+})
+
+app.get('/register', async (req, res) => {
+    const users = await User.find()
+
+    res.json(users)
+})
+
+
 
 app.listen(PORT, () => console.log(`Server started on port ${PORT}`))
