+++ src/pages/Navbar.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import './App.css'; 
+import '../App.css'; 
 import { useState } from "react";
 import { FaQuestionCircle, FaCog, FaToggleOn, FaToggleOff, FaWindowClose } from 'react-icons/fa'; 
 import Instructions from "./Instructions";
