+++ src/pages/WinPage.js
